requirements.txt
asgiref==3.1.2
auth0-python==3.6.1
boto3==1.9.146
botocore==1.12.146
channels==2.2.0
channels-redis==2.4.0
Django==2.1.7
django_compressor==2.2
django-dotenv==1.4.2
django-eventstream==2.5.0
django-extensions==2.1.6
django-filter==2.1.0
django-sass-processor==0.7.2
djangorestframework==3.9.1
djproxy==2.3.4
elasticsearch-dsl==6.3.1
gunicorn==19.9.0
Jinja2==2.10
kubernetes==8.0.1
libsass==0.17.0
misaka==2.1.1
model_mommy==1.6.0
mozilla-django-oidc==1.2.1
psycopg2-binary==2.7.7
pytest==4.3.0
pytest-django==3.4.7
pytest-spec==1.1.0
python-jose[cryptography]==3.0.1
raven==6.10.0
service_identity==18.1.0
urllib3<1.25
uvicorn==0.6.1
